app/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
pymysql
python-dotenv
cryptography
mysqlclient
flask-jwt-extended
flask-cors
google-cloud-secret-manager
